fix: refactor go client
